✨ 自分の登録だけ抹消できる機能に変更

+++ app.py
@@ -161,19 +161,24 @@ def dashboard():
                                 id=id)
 
 @app.route('/delete/<int:id>', methods=['GET', 'POST'])
+@login_required
 def delete(id):
-    user_to_delete = Users.query.get_or_404(id)
-    name = None
-    form = UserForm()
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash('User Deleted Successfully!')
-        our_users = Users.query.order_by(Users.date_added)
-        return render_template('add_user.html', form=form, name=name, our_users=our_users)
-    except:
-        flash('Whoops! There was a problem deleting user, try again...')
-        return render_template('add_user.html', form=form, name=name, our_users=our_users)
+    if id == current_user.id:
+        user_to_delete = Users.query.get_or_404(id)
+        name = None
+        form = UserForm()
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash('User Deleted Successfully!')
+            our_users = Users.query.order_by(Users.date_added)
+            return render_template('add_user.html', form=form, name=name, our_users=our_users)
+        except:
+            flash('Whoops! There was a problem deleting user, try again...')
+            return render_template('add_user.html', form=form, name=name, our_users=our_users)
+    else:
+        flash("Sorry, you can't delete that user!")
+        return redirect(url_for('dashboard'))
 
 @app.route('/posts/delete/<int:id>')
 @login_required
